refactor: update auth-related types and imports for consistency across services
