+++ client/src/types/Auth.ts
@@ -1,5 +1,19 @@
+import type { UserModel } from "./User";
+
+
 export interface RegisterForm {
     Username: string;
     Email: string;
     Password: string;
+}
+
+export interface LoginForm {
+    Email: string;
+    Password: string;
+}
+
+export interface AuthResponse {
+  User: UserModel;
+  Message?: string;
+  Token?: string;
 }
